requirement.txt
openai
bandit
pytest
python-dotenv